Add toolchain files for Cortex-M0/M4 cross-compilation
These toolchain files contain the necessary steps that enable cross-compile
for Cortex-M0/M4 targets using Clang or GCC as compilers.

Note: arm-none-eabi-gcc toolchain is needed for this to work.
The 'bin' subdirectory of the toolchain has to be in your PATH.

Usage: Add -DCMAKE_TOOLCHAIN_FILE=path/to/csibe/toolchain/toolchain.file
to cmake options.